feat: 복약 스케줄 자동 생성 및 복약 여부 체크 기능 추가 (#5)
* fix: 응답 누락 허용 횟수 필드 추가

* fix: sms 인증 swagger 설명 추가

* feat: medication scheduler 구현

* feat: 복약 여부 체크 및 날짜별 복약 스케줄 조회 기능 구현

* fix: 약 복용시간 중복 선택 가능하도록 수정

* fix: Medication CRUD 수정

* feat: Medication, MedicationSchedule 리팩토링 및 api 추가

+++ src/main/java/com/solution/Ongi/domain/medication/service/MedicationService.java
@@ -2,21 +2,24 @@ package com.solution.Ongi.domain.medication.service;
 
 import com.solution.Ongi.domain.medication.Medication;
 import com.solution.Ongi.domain.medication.dto.CreateMedicationRequest;
+import com.solution.Ongi.domain.medication.dto.MedicationDTO;
+import com.solution.Ongi.domain.medication.dto.UpdateMedicationRequest;
 import com.solution.Ongi.domain.medication.repository.MedicationRepository;
 import com.solution.Ongi.domain.user.User;
 import com.solution.Ongi.domain.user.repository.UserRepository;
 import com.solution.Ongi.domain.user.service.UserService;
-import jakarta.transaction.Transactional;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
+import com.solution.Ongi.global.response.code.ErrorStatus;
+import com.solution.Ongi.global.response.exception.GeneralException;
 import java.time.LocalTime;
 import java.time.format.DateTimeFormatter;
 import java.util.List;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 @Service
-@Transactional
 @RequiredArgsConstructor
+@Transactional(readOnly = true)
 public class MedicationService {
 
     private final UserRepository userRepository;
@@ -24,29 +27,64 @@ public class MedicationService {
     private final MedicationRepository medicationRepository;
     private final DateTimeFormatter timeFormatter=DateTimeFormatter.ofPattern("HH:mm");
 
-    //Meal 생성
-    public Medication createMedication(Long userId, CreateMedicationRequest createMedicationRequest){
-        User user=userService.getUserByIdOrThrow(userId);
+    // Medication 생성
+    @Transactional
+    public Medication createMedication(String userId, CreateMedicationRequest createMedicationRequest){
+        User user=userService.getUserByLoginIdOrThrow(userId);
 
         Medication medication=Medication.builder()
                 .medication_title(createMedicationRequest.getMedication_title())
-                .medication_time(LocalTime.parse(createMedicationRequest.getMedication_time(),timeFormatter))
+                .medication_time(
+                    createMedicationRequest.getTimeList().stream()
+                        .map(time -> LocalTime.parse(time, timeFormatter))
+                        .toList()
+                )
                 .user(user)
                 .build();
 
         return medicationRepository.save(medication);
     }
 
-    //유저의 Meal 전체 조회
-    public List<Medication> getAllMedication(Long userId){
-        userService.getUserByIdOrThrow(userId);
-        return medicationRepository.findByUserId(userId);
+    // 유저의 Medication 전체 조회
+    public List<MedicationDTO> getAllMedication(String loginId){
+        User user = userService.getUserByLoginIdOrThrow(loginId);
+        List<Medication> medications = medicationRepository.findAllByUserId(user.getId());
+        List<MedicationDTO> result = medications.stream()
+            .map(MedicationDTO::new)
+            .toList();
+        return result;
     }
 
-    //Meal 삭제
-    public void deleteMedication(Long medication_id){
+    // Medication 삭제
+    @Transactional
+    public void deleteMedication(String loginId, Long medication_id){
+        User user = userService.getUserByLoginIdOrThrow(loginId);
         Medication medication=medicationRepository.findById(medication_id)
-                .orElseThrow(()->new RuntimeException("약이 존재하지 않습니다."));
+                .orElseThrow(()->new GeneralException(ErrorStatus.MEDICATION_NOT_FOUND));
+
+        if (!medication.getUser().getId().equals(user.getId())){
+            throw new GeneralException(ErrorStatus.UNAUTHORIZED_ACCESS);
+        }
+
         medicationRepository.delete(medication);
     }
+
+    // Medication 수정
+    @Transactional
+    public void updateMedication(String loginId, Long medicationId, UpdateMedicationRequest request) {
+        User user = userService.getUserByLoginIdOrThrow(loginId);
+        Medication medication = medicationRepository.findById(medicationId)
+            .orElseThrow(() -> new GeneralException(ErrorStatus.MEDICATION_NOT_FOUND));
+
+        if (!medication.getUser().getId().equals(user.getId())) {
+            throw new GeneralException(ErrorStatus.UNAUTHORIZED_ACCESS);
+        }
+
+        List<LocalTime> convertedTimeList = request.timeList().stream()
+            .map(LocalTime::parse) // "12:30" → LocalTime
+            .toList();
+
+        medication.update(request.title(), convertedTimeList);
+    }
+
 }
